# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Storage and Send API\nFirst login with user and use authorization as \"Bearer JWTTOKEN\"",
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/send": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Send request with bind id or name",
                "consumes": [
                    "text/plain"
                ],
                "tags": [
                    "run"
                ],
                "summary": "Send run the control; methods depending in control",
                "parameters": [
                    {
                        "type": "string",
                        "description": "set endpoint",
                        "name": "endpoint",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "set control",
                        "name": "control",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "send key values",
                        "name": "payload",
                        "in": "body",
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "respond from related server",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Send request with bind id or name",
                "consumes": [
                    "text/plain"
                ],
                "tags": [
                    "run"
                ],
                "summary": "Send run the control; methods depending in control",
                "parameters": [
                    {
                        "type": "string",
                        "description": "set endpoint",
                        "name": "endpoint",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "set control",
                        "name": "control",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "send key values",
                        "name": "payload",
                        "in": "body",
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "respond from related server",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apimodels.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apimodels.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "some problem"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}